requirements.txt
Django==1.8
django-crispy-forms==1.6.0
django-filter==0.13.0
django-guardian==1.4.5
djangorestframework==3.4.3
ecdsa==0.13
Fabric==1.12.0
Markdown==2.6.6
MySQL-python==1.2.5
paramiko==1.17.2
pbr==1.10.0
pycrypto==2.6.1
six==1.10.0
stevedore==1.16.0
virtualenv==15.0.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
